pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hackathon-2025-3to1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.2",
    "ipykernel>=7.1.0",
    "openai>=2.8.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.38.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
